1.2 Added Construction Worker
Implemented work and commute in the construction worker where work is only implemented in construction worker not adult, because construction worker defines what kind of work they do.
Commute is implemented in Adult making it a virtual not an abstract because there needs to be a standard implementation of commute, since not every kind of profession will have a unique commute method.